printf finished

+++ C/RANK 2/ft_printf/specifiers/u_specifier.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   u_specifier.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: llopes-d <llopes-d@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/04/30 16:29:58 by llopes-d          #+#    #+#             */
+/*   Updated: 2023/04/30 16:30:41 by llopes-d         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../ft_printf.h"
 
-static int ft_putstr_u(char *str)
+static int	ft_putstr_u(char *str)
 {
-	int index;
+	int	index;
 
 	index = 0;
 	while (str[index])
@@ -13,9 +25,9 @@ static int ft_putstr_u(char *str)
 	return (index);
 }
 
-static int get_length_u(long int i)
+static int	get_length_u(long int i)
 {
-	int length;
+	int	length;
 
 	length = 0;
 	if (i == 0)
@@ -33,11 +45,11 @@ static int get_length_u(long int i)
 	return (length);
 }
 
-int u_specifier(long int nb)
+int	u_specifier(unsigned int nb)
 {
-	char str[50];
-	int length;
-	long int unb;
+	char		str[50];
+	int			length;
+	long int	unb;
 
 	unb = nb;
 	length = get_length_u(unb);
